+++ eventum/ui/src/pages/ProjectPage/OutputPluginsTab/OutputPluginParams/index.tsx
@@ -5,6 +5,7 @@ import { FileOutputPluginParams } from './FileOutputPluginParams';
 import { HTTPOutputPluginParams } from './HTTPOutputPluginParams';
 import { KafkaOutputPluginParams } from './KafkaOutputPluginParams';
 import { OpensearchOutputPluginParams } from './OpensearchOutputPluginParams';
+import { S3OutputPluginParams } from './S3OutputPluginParams';
 import { StdoutOutputPluginParams } from './StdoutOutputPluginParams';
 import { TcpOutputPluginParams } from './TcpOutputPluginParams';
 import { UdpOutputPluginParams } from './UdpOutputPluginParams';
@@ -32,6 +33,7 @@ const pluginNamesToParamsComponent = {
   http: HTTPOutputPluginParams,
   kafka: KafkaOutputPluginParams,
   opensearch: OpensearchOutputPluginParams,
+  s3: S3OutputPluginParams,
   stdout: StdoutOutputPluginParams,
   tcp: TcpOutputPluginParams,
   udp: UdpOutputPluginParams,